% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{Bt_prob_abu}
\alias{Bt_prob_abu}
\title{Bt_prob_abu(x) is a function of estimating the species probabilities in the bootstrap assemblage based on abundance data.}
\usage{
Bt_prob_abu(x)
}
\arguments{
\item{x}{a vector of species sample frequencies.}
}
\value{
a numeric vector.
}
\description{
Bt_prob_abu(x) is a function of estimating the species probabilities in the bootstrap assemblage based on abundance data.
}
