% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{Chao_Hill_abu}
\alias{Chao_Hill_abu}
\title{Chao_Hill_abu(x, q) is a function of obtaining estimators of Hill numbers of order q based on abundance data.}
\usage{
Chao_Hill_abu(x, q)
}
\arguments{
\item{x}{a vector of species sample frequencies.}

\item{q}{a numeric or a vector of diversity order; The suggested range for q is [0, 3].}
}
\value{
a numerical vector of diversity.
}
\description{
Chao_Hill_abu(x, q) is a function of obtaining estimators of Hill numbers of order q based on abundance data.
}
