% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{ChaoHill}
\alias{ChaoHill}
\title{ChaoHill(dat, datatype, from, to, interval, B, conf) is the function of calculating the empirical and the proposed diversity profile,
their bootsrap standard errors and confidance intervals.}
\usage{
ChaoHill(
  dat,
  datatype = c("abundance", "incidence"),
  from = 0,
  to = 3,
  interval = 0.1,
  B = 1000,
  conf = 0.95
)
}
\arguments{
\item{dat}{a vector of species sample frequencies (for abundance data), or incidence-based sample frequencies (1st entry must be the number of sampling unit).}

\item{datatype}{a character of data type,"abundance" or "incidence".}

\item{from}{a numeric number of diversity order q (the start order of profile).}

\item{to}{a numeric number of diversity order q (the end order of profile).}

\item{interval}{a numeric number to specify each increment of q from the start to end order.}

\item{B}{an integer to specify the number of bootstrap replications, B = 1000 is suggested for constructing confidence intervals;
To save running time, use a smaller value (e.g. B = 200).}

\item{conf}{a confidence coefficient between 0 and 1.}
}
\value{
a list, consisting of 4 matrices including respectively diversity estimates, bootstrap standard errors, lower confidence bounds, and upper confidence bounds.
In each matrix, the first row gives the results for the empirical diversity, and the second row gives the results for the proposed diversity estimates.
Columns give the results for different orders of q.
}
\description{
ChaoHill(dat, datatype, from, to, interval, B, conf) is the function of calculating the empirical and the proposed diversity profile,
their bootsrap standard errors and confidance intervals.
}
