% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSpadeR.R
\name{Chat.Ind}
\alias{Chat.Ind}
\title{Estimate sample coverage}
\usage{
Chat.Ind(ab, m = sum(ab))
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{m}{Scalar, sample size at which to estimate coverage.}
}
\value{
Scalar between 0 and 1, estimated sample coverage.
}
\description{
Computes Chao and Jost 2012's estimate of sample coverage, assuming that
individuals are independently and randomly sampled. This function computes
rarefaction and extrapolation estimates of coverage given an integer sample
size \code{m}. We modified the source code to default to observed sample size
for convenience, and so that parameter names matched analogous use elsewhere
in the \code{MeanRarity} package. This function is copied directly from
\code{SpadeR:::Chat.Ind()} from the R package SpadeR 0.1.1 by Anne Chao, K. H. Ma,
T. C. Hsieh and Chun-Huo Chiu.
}
\examples{

# generate sample
abs <- sample_infinite(fit_SAD(rich = 50, simpson = 20)[[3]], 150)

# estimate coverage of sample
Chat.Ind(abs)

# estimate coverage under rarefaction
Chat.Ind(abs, m = 100)
}
\references{
\insertRef{Chao2012a}{MeanRarity}
}
