% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{checkchao}
\alias{checkchao}
\title{Assess Chao and Jost 2015 Hill diversity CI}
\usage{
checkchao(x, B, l, truediv, conf = 0.95)
}
\description{
Given empirical abundances \code{x}, use the Chao and Jost 2015 MEE method
to estimate the sampling distribution of the Hill diversity (given exponent
\code{l}), and see where known, true diversity \code{truediv} falls in that
distribution
}
\details{
This function could be extended to look at e.g. incidence-based estimations.

@param x Numeric vector of integer species abundances in a sample
@param B Scalar, number of replicat bootstrap draws
@param l Scalar, exponent for scaling rarity in computing Hill diversity
@param truediv Scalar, known true Hill diversity of the pool from whcih
sample is drawn, for comparison to estimated sampling distribution
@param conf Scalar, target coverage probability of estimated CI

@return data.frame with estimated p-value for true diversity, the diversity
values of the upper and lower estimated cofidence limits, the asymptotic
Hill diversity estimate, and the empirical diversity

@export
}
