% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compare_ests}
\alias{compare_ests}
\title{Estimotor accuracy and bias}
\format{
A data.frame with 7919208 rows and 9 variables:
\itemize{
\item \emph{truth}: Numeric, diversity of simulated SAD
\item \emph{chaoest} Numeric, Chao and Jost 2015 asymptotic diversity estimate
\item \emph{naive}: Numeric, Hill diversity of sample frequencies
\item \emph{gods}: Numeric, God's estimator of true Hill diversity
\item \emph{n}: Numeric, individuals sampled
\item \emph{ell}: Scalar, scaling parameter for Hill diversity
\item \emph{SAD}: index of SAD sampled
\item \emph{distribution} Factor, family, one of "lnorm" or "gamma"
\item \emph{fitted.parameter} Factor, shape parameter for fitting SAD
}
}
\usage{
compare_ests
}
\description{
A simulated dataset containing estimated Hill diversity based on finite
samples from simulated species abundance distributions. Hill diversity is
estimated with the naïve estimator, \code{rarity(ab, l)}, the Chao and Jost
2015 non-parametric asymptotic estimator \code{Chao_Hill_abu(ab, l)}, and God's
estimator \code{GUE(sample_freq, true_freq, l)}.
}
\keyword{datasets}
