% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{conf.reg}
\alias{conf.reg}
\title{conf.reg(x_axis,LCL,UCL,...) is a function to plot the confidence region.}
\usage{
conf.reg(x_axis, LCL, UCL, ...)
}
\arguments{
\item{x_axis}{a vector of diversity orders.}

\item{LCL}{a vector of lower confidence bounds.}

\item{UCL}{a vector of upper confidence bounds.}

\item{...}{further arguments to be passed to \code{polygon}}
}
\value{
a polygon plot
}
\description{
\code{conf.reg} uses polygon to draw a confidence band plot
}
