% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot_gamma_and_lnorm.R
\name{divers_lnorm}
\alias{divers_lnorm}
\title{relative abundances given lognormal disribution}
\usage{
divers_lnorm(rich, x)
}
\arguments{
\item{rich}{Total number of species in the SAD, an integer}

\item{x}{Shape paramter sdlog, a scalar}
}
\description{
This is a wrapper for `qlnrom` that
takes the number of species and a shape parameter (sd of the log(lognormal)))
and gives relative abundance estimates for each species
}
\examples{
divers_lnorm(rich = 10, x = 1)
}
\seealso{
qlnorm
}
