% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFuns.R
\name{e2d}
\alias{e2d}
\title{Hill diversity from evenness}
\usage{
e2d(e, rich)
}
\arguments{
\item{e}{Scalar, evenness value between 0 and 1}

\item{rich}{Integer, species richness}
}
\description{
returns Hill diversity based on evenness and richness
}
\details{
This function is the inverse of e3Fun
}
\seealso{
e3Fun
}
