% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFuns.R
\name{e3Fun}
\alias{e3Fun}
\title{Normalized Hill evenness}
\usage{
e3Fun(ab, l, q = NULL)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{l}{Scaling exponent for the mean, can be any real number.}

\item{q}{Scalar, traditional Hill number scaling exponent, q = 1-l where l is the
scaling parameter for the generalized mean. Can be any real number.}
}
\description{
Computes assemblage evenness measure based on the ratio of Hill diversities.
}
\details{
There are many valid ways to compute evenness, and this one is the
"normalized slope" of the Hill diversity profile. The normalization means
that for any richness, the maximum evenness is 1, and the minimum, achieved
when 1 species has practically all of the relative abundance in the
assemblage, is 0. This evenness metric was defined by Chao and Ricotta
(2019) as measure "E3". Like Hill diversity itself, evenness is considered a
function of scale, as well as the relative abundances in the assemblage.
}
\seealso{
e2d
}
\concept{Computation}
