% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effort_rare}
\alias{effort_rare}
\title{Rarefied diversity estimates for each site, multiple standards}
\format{
A tibble with 1008 rows and 6 variables:
\itemize{
\item \emph{site}: Factor, site ID
\item \emph{divind}: Factor, Hill diversity ("richness"
, "Hill-Shannon", or "Hill-Simpson)
\item \emph{etype} Factor, estimate type ("observed", or "asymptotic")
\item \emph{diversity}: Numeric, mean estimated Hill diversity value
\item \emph{method}: Factor, standardization method ("effort", "size", or "coverage")
\item \emph{xax}: Numeric, number of transects in subsamples, mean number of
individuals in subsamples, or mean sample coverage of subsamples
\item \emph{obsshan}: Numeric, mean observed Hill-Shannon diversity
\item \emph{chaoshan} Numeric, mean estimated asymptotic Hill-Shannon diversity
\item \emph{obssimp} Numeric, mean observed Hill-Simpson diversity
\item \emph{chaosimp} Numeric, mean estimated asymptotic Hill-Simpson diversity
}
}
\usage{
effort_rare
}
\description{
?
}
\seealso{
\insertRef{Chao2012}{MeanRarity}.
}
\keyword{datasets}
