% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{ellnotate}
\alias{ellnotate}
\title{Convenience function to label rarity plots with scaling exponent}
\usage{
ellnotate(title = NULL, l, unicode_in_title = TRUE, ...)
}
\arguments{
\item{title}{Character string, or NULL}

\item{l}{Scaling exponent for the mean, can be any real number.}

\item{unicode_in_title}{Logical, include unicode ell and its value}
}
\value{
ggplot object
}
\description{
Convenience function to label rarity plots with scaling exponent
}
\examples{
myp <- rarity_plot(1:10, 1)
myp +
 ellnotate("Hill-Shannon diversity", l = 1)

}
\concept{Visualization}
