% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaubyHardy.R
\name{ENSer}
\alias{ENSer}
\title{Worker to iteratively solve ENS transformation of Hurlbert diversity}
\usage{
ENSer(s_k, ens_k_old, k)
}
\arguments{
\item{s_k}{numeric Hurlbert rarefaction}

\item{ens_k_old}{numeric ENS estimate from previous iteration}

\item{k}{integer sample size parameter for rarefaction}
}
\value{
numeric scalar, updated ENS estimate
}
\description{
Worker to iteratively solve ENS transformation of Hurlbert diversity
}
