% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{err_plot_data}
\alias{err_plot_data}
\title{Sampling variability in Hill diversity}
\format{
A data.frame with 7224 rows and 9 variables:
\itemize{
\item \emph{evenness} Factor, approximately the ratio of Hill-Simpson to richness
\item \emph{distribution} Factor, family, one of "lnorm" or "gamma"
\item \emph{ell}: Scalar, scaling parameter for Hill diversity
\item \emph{sample_sdlog}: Numeric, standard deviation of natural log of sample
Hill diversity
\item \emph{estimator_sdlog}: Numeric, standard deviation of natural log of
estimated Hill diversity of the SAD based on asymptotic estimator
\item \emph{sample_bias} Numeric, mean log ratio of sample diversity to expected
sample diversity
\item \emph{naive_bias} Numeric, mean log ratio of sample diversity to true diversity
\item \emph{estimator_bias} Numeric, mean log ratio of estimated asymptotic diversity
to true diversity
\item \emph{sample_rmsle} Numeric, square root of the mean squared log ratio of
sample diversity to expected sample diversity
\item \emph{estimator_rmsle} Numeric, square root of the mean squared log ratio of
estimated asymptotic diversity to true diversity
\item \emph{estimator_rmsle} Numeric, square root of the mean squared log ratio of
estimated asymptotic diversity to true diversity
\item \emph{true_diversity} Numeric, Hill diversity of the SAD
\item \emph{mean_sample} Numeric, mean of sample diversity for combination of ell,
evenness, distribution, and SS.
\item \emph{mean_asymptotic} Numeric, mean of estimated asymptotic diversity for
combination of ell, evenness, distribution, and SS.
\item \emph{sample_size} Integer, number of individuals sampled
}
}
\usage{
err_plot_data
}
\description{
A simulated dataset with mean absolute error and log sd for sample and
asymptotic Hill diversity estimates based on finite samples from simulated
species abundance distributions. Hill diversity was estimated with the naïve
estimator, \verb{rarity(sample_freq, l) and the Chao and Jost 2015 non-parametric asymptotic estimator }Chao_Hill_abu(sample_freq, l), and parameterized by
richness and evenness
}
\references{
\insertRef{Chao2015}{MeanRarity}
\insertRef{Chao2019a}{MeanRarity}
}
\keyword{datasets}
