% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{errs}
\alias{errs}
\title{Estimator accuracy and bias}
\format{
A data.frame with 2376 rows and 8 variables:
\itemize{
\item \emph{ell}: Scalar, scaling parameter for Hill diversity
\item \emph{distribution} Factor, family, one of "lnorm" or "gamma"
\item \emph{fitted.parameter} Factor, shape parameter for fitting SAD
\item \emph{n}: Numeric, individuals sampled
\item \emph{SAD}: Integer, index of SAD sampled
\item \emph{estimator} Character, estimator whose error is described
\item \emph{rmse} Numeric, root mean squared error of estimator
\item \emph{bias} Numeric, mean difference between estimator and true diversity
}
}
\usage{
errs
}
\description{
A simulated dataset with rmse and mean absolute error for different
approaches to estimating the true (asymptotic) Hill diversity based on finite
samples from simulated species abundance distributions. Hill diversity was
estimated with the naïve estimator, \code{rarity(sample_freq, l)}, the Chao and
Jost 2015 non-parametric asymptotic estimator \code{Chao_Hill_abu(sample_freq, l)}, and God's estimator \code{GUE(sample_freq, true_freq, l)}.
}
\references{
\insertRef{Chao2015}{MeanRarity}
}
\keyword{datasets}
