% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{Exp}
\alias{Exp}
\title{Exponentiate \code{x} to power \code{y}}
\usage{
Exp(x, y)
}
\arguments{
\item{x}{Numeric vector or scalar}

\item{y}{Numeric scalar}
}
\value{

}
\description{
Exponentiate \code{x} to power \code{y}
}
\examples{

all.equal(Exp(exp(1), 5),  exp(5) )
Ln(Exp(exp(1), 10)) ==10
}
