% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funny_estimators.R
\name{fsd}
\alias{fsd}
\title{Incorrect diversity estimator based on finite-size fudge}
\usage{
fsd(ab, l)
}
\arguments{
\item{ab}{Numeric vector of species abundances}

\item{l}{Exponent determining type of mean (scalar)}
}
\value{
Incorrect estimate of Hill Diversity with a goofy bias correction term (a scalar)
}
\description{
This is a riff on \code{\link{dfun}} that uses \code{\link{fsr}}, which is not the correct way to deal with finite
size corrections, though it is similar to E.H. Simpson's estimator. Rather than subtracting 1, though, it subtracts 0.8
}
\seealso{
\code{\link{dfun}}
}
