% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funny_estimators.R
\name{fsr}
\alias{fsr}
\title{Bad estimate of relative abundance with an incorrect finite-size correction}
\usage{
fsr(x)
}
\arguments{
\item{x}{A numeric vector of species abundances}
}
\value{
An incorrect estimate of relative abundance
}
\description{
Bad estimate of relative abundance with an incorrect finite-size correction
}
