% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{GUE}
\alias{GUE}
\title{God's unbiased estimator}
\usage{
GUE(freqs, true_p, l)
}
\arguments{
\item{freqs}{Non-negative numeric vector of observed species frequencies}

\item{true_p}{Non-negative numeric vector of true, population-level species
frequencies}

\item{l}{Scaling exponent for the mean, can be any real number.}
}
\description{
The mean rarity perspective suggests a novel approach to estimating true
diversity from a sample, using the fact that the sample mean is an unbiased
estimate of the population mean. Although how to estimate a species’ rarity
from sample data remains unsolved, we test the promise of this idea using
what we call God’s estimator: All-knowing, God can plug in the true
population parameters for each species. Nevertheless, maybe for fun, God
chooses to compute the sample mean, weighting species’ true rarities by
their sampled frequency.
}
\examples{
GUE(freqs =  1:10/sum(1:10), true_p =  1:10/sum(1:10), l = 1) # true richness 10
GUE(freqs =  rep(1:5,2)/sum(1:10), true_p =  1:10/sum(1:10), l = 1) # true richness 10
}
