% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{Hill}
\alias{Hill}
\title{Hill(x, q, datatype) is a function of obtaining the empirical Hill numbers of order q based on abundance data or incidence data.}
\usage{
Hill(x, q, datatype = c("abundance", "incidence"))
}
\arguments{
\item{x}{a vector of species sample frequencies (for abundance data), or incidence-based sample frequencies (1st entry must be the number of sampling unit).}

\item{q}{a numeric or a vector of diversity order.}

\item{datatype}{a character of data type,"abundance" or "incidence".}
}
\value{
a numerical vector.
}
\description{
Hill(x, q, datatype) is a function of obtaining the empirical Hill numbers of order q based on abundance data or incidence data.
}
