% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaubyHardy.R
\name{hRare}
\alias{hRare}
\title{Hurlbert rarefaction}
\usage{
hRare(ab, k)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{k}{integer sample size parameter for rarefaction}
}
\value{
numeric scalar, unbiased estimate of rarefied richness, assuming
abundances are sampled without replacement from a finite species pool
}
\description{
Provides Hurlberts (1971) rarefied richness.
}
\examples{
hRare(1:10, 2)
}
\seealso{
vegan::rarefy()
}
\concept{Computation}
