% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaubyHardy.R
\name{hurl}
\alias{hurl}
\title{Estimate ENS based on Hurlbert rarefaction}
\usage{
hurl(ab, k, maxit = 1e+05, tol = 1e-12)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{k}{integer sample size parameter for rarefaction}

\item{maxit}{integer, maximum number of iterations}

\item{tol}{numeric, threshold for convergence}
}
\value{
Numeric scalar: estimated Hurlbert ENS
}
\description{
Implements estimate described in Dauby and Hardy 2011 for a class of
rarefaction-based ENS diversity estimates. These estimates suffer from
minimal bias and are quite efficient, while retaining some of the nice
properties of Hill diversity metrics. They are parameterized by sample size
\code{k}, and when \code{k == 2} they are equivalent to Hill-Simpson diversity. One
interpretation is that this ENS is the number of species in a perfectly even
assemblage that would have the same rarefied richness as the focal
assemblage/sample. Larger k values emphasize rare species, and as k
approaches community size the Hulbert ENS approaches true richness. Unbiased
estimators are given for k < sample size.
}
\examples{
ab = sample(10:50, 50, replace =TRUE)
hurl(ab, 2)
# not run
# hurl(ab, 1e5) # returns an error

}
\references{
\insertRef{Dauby2012}{MeanRarity}
\insertRef{Hurlbert1971}{MeanRarity}
}
\concept{Estimators}
