% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFuns.R
\name{ipfun}
\alias{ipfun}
\title{Inverse function for power transformation.}
\usage{
ipfun(x, pow)
}
\arguments{
\item{x}{A numeric vector.}

\item{pow}{The reciprocal of the exponent to which all values
in \code{x} are raised.}
}
\value{
A numeric vector with same length as \code{x}.
}
\description{
Transform a vector by raising to a power... specifically 1/pow.
}
