% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{Ln}
\alias{Ln}
\title{Robust natural log}
\usage{
Ln(x)
}
\arguments{
\item{x}{Numeric vector or scalar}
}
\value{
natural log transformation of \code{x}, \code{Ln(0) ==0}
}
\description{
Robust natural log
}
\examples{
Ln(5)== log(5)
Ln(0)
# base log(0) returns unhelpful -Inf
log(0)

}
