% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_ests}
\alias{mean_ests}
\title{Rarefied diversity estimates for each site}
\format{
A tibble with 2015 rows and 8 variables:
\itemize{
\item \emph{site}: Character, site ID
\item \emph{individuals}: Integer, individuals in rarefied sample
\item \emph{obsrich}: Numeric, mean observed species richness
\item \emph{chaorich}: Numeric, mean estimated asymptotic richness from Chao1
estimator
\item \emph{obsshan}: Numeric, mean observed Hill-Shannon diversity
\item \emph{chaoshan} Numeric, mean estimated asymptotic Hill-Shannon diversity
\item \emph{obssimp} Numeric, mean observed Hill-Simpson diversity
\item \emph{chaosimp} Numeric, mean estimated asymptotic Hill-Simpson diversity
}
}
\usage{
mean_ests
}
\description{
A dataset containing expected observed Hill diversity and estimated
asymptotic Hill diversity for each of the 4 community samples analyzed in
\insertCite{Roswell2020}{MeanRarity}.
}
\references{
\insertRef{Roswell2020}{MeanRarity}
}
\keyword{datasets}
