% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{mult_trans}
\alias{mult_trans}
\title{Multiplicative transformation}
\usage{
mult_trans()
}
\description{
This transformation provides symmetrical-looking multiplicative factors from
log-scaled variables
}
\examples{
plot(mult_trans(), xlim = c(log(0.1), log(10)))
}
