% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFuns.R
\name{obs_est}
\alias{obs_est}
\title{Observed and asymptotic diversity}
\usage{
obs_est(ab)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}
}
\value{
Dataframe with total abundance and observed and asymptotic diversity
estimates for the Pythagorean mean rarities: richness, exponentiated Shannon,
and inverse Simpson.
}
\description{
Computes observed and asymptotic Hill diversity estimates from a vector of
integer abundances
}
\details{
Only integer abundances allowed in \code{ab}, as asymptotic estimator relies
on sampling theory for individuals.
}
\examples{
obs_est(c(20,8,5,4,2,1))
}
\seealso{
\code{\link{rarity}}, \code{\link[SpadeR]{Choa_Hill_Abu}}
}
