% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{omit_y}
\alias{omit_y}
\title{Convenience function to omit y-axis elements for constructing multi-panel plots.}
\usage{
omit_y(p)
}
\arguments{
\item{p}{A ggplot.}
}
\value{
A ggplot.
}
\description{
This function removes, y-axis elements for plots that can
be included (klugely) in a multi-panel plot. In this case the plotted area varies.
}
\seealso{
\code{\link{white_y}}
}
