% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFuns.R, R/pfun.R
\name{pfun}
\alias{pfun}
\title{Power transformation}
\usage{
pfun(x, pow)

pfun(x, pow)
}
\arguments{
\item{x}{A numeric vector.}

\item{pow}{The value of the exponent to which all values are raised.}
}
\value{
A numeric vector with same length as \code{x}.

A numeric vector with same length as \code{x}.
}
\description{
Transform a vector by raising to a power.

Transform a vector by raising to a power.
}
