% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_data}
\alias{plot_data}
\title{Sampling variability in Hill diversity}
\format{
A data.frame with 7224 rows and 9 variables:
\itemize{
\item \emph{evenness} Factor, approximately the ratio of Hill-Simpson to richness
\item \emph{distribution} Factor, family, one of "lnorm" or "gamma"
\item \emph{ell}: Scalar, scaling parameter for Hill diversity
\item \emph{SS} Integer, number of individuals sampled
\item \emph{sample}: Numeric, Hill diversity of the sample
\item \emph{asymptotic}: Numeric, estimated Hill diversity of the SAD based on
asymptotic estimator
\item \emph{true_diversity} Numeric, Hill diversity of the SAD
\item \emph{mean_sample} Numeric, mean of sample diversity for combination of ell,
evenness, distribution, and SS.
\item \emph{mean_asymptotic} Numeric, mean of estimated asymptotic diversity for
combination of ell, evenness, distribution, and SS.
}
}
\usage{
plot_data
}
\description{
A simulated dataset with mean absolute error and log sd for sample and
asymptotic Hill diversity estimates based on finite samples from simulated
species abundance distributions. Hill diversity was estimated with the naïve
estimator, \verb{rarity(sample_freq, l) and the Chao and Jost 2015 non-parametric asymptotic estimator }Chao_Hill_abu(sample_freq, l), and parameterized by
richness and evenness
}
\references{
\insertRef{Chao2015}{MeanRarity}
\insertRef{Chao2019}{MeanRarity}
}
\keyword{datasets}
