% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{power_trans}
\alias{power_trans}
\title{scale transformation based on \code{pfun} and \code{ipfun}}
\usage{
power_trans(pow, nb)
}
\arguments{
\item{pow}{exponent of power tranformation, scalar}

\item{nb}{number of desired breaks (approximate), scalar}
}
\value{
a scale transformation object for plotting in ggplot
}
\description{
This is a function for nice scale transformations for \code{ggplot2}, including when \code{pow < 0}
}
\seealso{
\code{scales::trans_new, scales::transbreaks, pfun, ipfun, prettify}
}
