% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{prettify}
\alias{prettify}
\title{Round numbers, more aggressively the larger they are}
\usage{
prettify(breaks)
}
\arguments{
\item{breaks}{is a numeric vector
intended for the breaks returned by \code{scales::transbreaks}}
}
\value{
numeric vector of same length as input, rounded nicely
}
\description{
Round numbers, more aggressively the larger they are
}
\examples{
prettify(sqrt(2:7))
prettify(sqrt(seq(2e3, 2e5, 4739)))
}
