% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_scales.R
\name{print_operator}
\alias{print_operator}
\title{Ratio labels}
\usage{
print_operator(x, base = exp(1))
}
\arguments{
\item{x}{Numeric, on a logarithmic scale}

\item{base}{Scalar, base of the logarithm used}
}
\value{
String to be evaluated with \code{str2expression}
}
\description{
Ratio labels
}
\examples{
print_operator(((-1:3)))

}
\concept{Visualization}
