% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{radplot}
\alias{radplot}
\title{Plot a rank-abundance distribution}
\usage{
radplot(
  ab,
  maxrich = length(comm),
  maxab = max(comm),
  fill = "red",
  shape = 16,
  Whittaker = FALSE
)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{maxrich}{Scalar, how many species to include space for}

\item{maxab}{Scalar, how big is the largest abundance value}

\item{fill}{Character string specifying a color, could be extended to
permit name for community ID variable}

\item{shape}{Character string for point shape, could be extended to
permit name for community ID variable}

\item{Whittaker}{Logical, if \code{TRUE} log-transform abundances}
}
\value{
ggplot object with a rank-abundance plot
}
\description{
Take a vector of species [relative] abundances and plot a rank-abundance
distribution or Whittaker plot.
}
\examples{
radplot(c(20,8,5,4,2,1))
radplot(c(20,8,5,4,2,1), Whittaker = TRUE)
}
