% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFuns.R
\name{rarity}
\alias{rarity}
\title{Compute Hill diversity: the mean species rarity}
\usage{
rarity(ab, l, q = NULL)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{l}{Scaling exponent for the mean, can be any real number.}

\item{q}{Scalar, traditional Hill number scaling exponent, q = 1-l where l is the
scaling parameter for the generalized mean. Can be any real number.}
}
\value{
Generalized mean community rarity with scaling exponent \code{"l"}.

When \code{l = 1}, arithmetic mean rarity (species richness).

When \code{l = 0}, geometric mean rarity (Hill-Shannon diversity), Shannon's
entropy \insertCite{Shannon1963}{MeanRarity} exponentiated.

When \code{l = -1}, harmonic mean rarity (Hill-Simpson diversity),
the inverse of the Simpson concentration
\insertCite{Simpson1949}{MeanRarity}.
}
\description{
Compute the empirical Hill diversity from abundances or relative abundances.
Hill diversity is also the mean species rarity.
}
\details{
We parameterize Hill diversity \eqn{D} as a the frequency-weighted mean
species rarity, with scaling exponent l \deqn{D = \sum{p_i *
r_i^{\ell}}^{-\ell}} where rarity of species i \eqn{r_1 = 1/p_i}. When
\eqn{\ell = 0} this is defined base on the limit from the left and the right,
which is the geometric mean \deqn{\exp(\frac{\sum{p_i * \ln(r_i)}}{\sum{p_i}})}

This is equivalent to the \eqn{q} notation of Jost 2006
\deqn{D=\sum{p_i^q}^{\frac{1}{1-q}}}
where \eqn{q=1-l}.

This function can also be called with \code{dfun()}
}
\examples{
rarity(c(20,8,5,4,2,1), 1) #species richness
rarity(c(20,8,5,4,2,1), 0) # Hill-Shannon diversity
rarity(c(20,8,5,4,2,1), -1) # Hill-Simpson diversity
rarity(c(20,8,5,4,2,1), q = 2) # The parameter `q` can be used instead for
# traditional Hill number parameterization
}
\references{
\insertRef{Simpson1949}{MeanRarity}
\insertRef{Shannon1963}{MeanRarity}
}
\seealso{
\code{\link{pfun}}, \code{\link{ipfun}}
}
\concept{Computation}
