% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{rarity_plot}
\alias{rarity_plot}
\title{Construct the full balance plot for scale ell, with reference means=means}
\usage{
rarity_plot(ab, ell, means = -1:1, noco = 1, lines = F, ...)
}
\description{
Construct the full balance plot for scale ell, with reference means=means
}
\seealso{
This function depends on internal functions in the \code{MeanRarity}
package which can be accessed with \code{:::} e.g.
\code{MeanRarity:::scale_plot}
}
