% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{rarity_plot_b}
\alias{rarity_plot_b}
\title{Generate rarity plot noting \ell value and optional banner title}
\usage{
rarity_plot_b(
  ab,
  l,
  q = NULL,
  means = -1:1,
  noco = 1,
  lines = FALSE,
  l_banner = TRUE,
  title = NULL,
  ...
)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{l}{Scaling exponent for the mean, can be any real number.}

\item{q}{Scalar, traditional Hill number scaling exponent, q = 1-l where l is the
scaling parameter for the generalized mean. Can be any real number.}

\item{means}{Numeric vector of scaling exponent values corresponding to
reference (by default, Pythagorean) means.}

\item{noco}{Scalar, shrinks text and points if plotting multiple balance
plots in a single plotting window.}

\item{lines}{Logical, should each individual be plotted as a "box" or should
individuals be summarized simply as the height of a line segment.}

\item{l_banner}{Logical, include unicode \ell and \ell value?}

\item{title}{Character vector, optional banner title}

\item{...}{Additional arguments passed to other functions.}
}
\value{
ggplot object, a rarity plot with banner annotations
}
\description{
Generate rarity plot noting \ell value and optional banner title
}
\examples{
rarity_plot_b(ab = 1:10, 1)
}
