% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_scales.R
\name{rat_breaks}
\alias{rat_breaks}
\title{Compute breaks for ratio scale}
\usage{
rat_breaks(base = exp(1), n = 5)
}
\arguments{
\item{base}{Scalar, base of the logarithm in use (not implemented)}

\item{n}{scalar, target number of breaks (not implemented)}
}
\value{
Function to apply over a vector values to generate axis breaks
}
\description{
Function to compute tick marks evenly spaced on the log scale but with pretty
numbers on the ratio scale
}
\examples{
ggplot2::ggplot(data = data.frame(x= 1:6, y = seq(-1, 1.5, 0.5))
       , ggplot2::aes(x, y))+
             ggplot2::geom_point()+
             ggplot2::scale_y_continuous(
             , breaks = rat_breaks()
             ) +
             ggplot2::geom_hline(yintercept = 0, size = 0.2)
}
