% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{rmsle}
\alias{rmsle}
\title{Root mean squared log error}
\usage{
rmsle(x, true_x = mean(Ln(x)))
}
\arguments{
\item{x}{positive numeric vector}

\item{true_x}{numeric scalar, target x value, \code{mean(Ln(x))} by default}
}
\value{
scalar, root mean of the squared log ratio of observations to target value
}
\description{
Root mean squared log error
}
\examples{

x <- rlnorm(n = 1e5)
rmsle(x) # should be very close to sdlog of lnorm, 1
y <- rnorm(n = 5000, mean = 12 )
rmsle(y, true_x = Ln(12))
1/12
}
