% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFuns.R
\name{sample_infinite}
\alias{sample_infinite}
\title{Sample a \link{relative} abundance vector with replacement}
\usage{
sample_infinite(commSAD, size)
}
\description{
@param commSAD Numeric vector of \link{relative} abundances
@param size Scalar, number of individuals in finite samples
}
\details{
@return A vector of integer species abundances

@export
}
