% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funny_estimators.R
\name{sApp}
\alias{sApp}
\title{Simpson's estimator of Hill-Simpson Diversity (we think?)}
\usage{
sApp(ab)
}
\arguments{
\item{ab}{Numeric vector of species abundances}
}
\value{
the reciprocal of the Simpson's concentration, a low-bias estimator for Hill-Simpson diversity (scalar)
}
\description{
E.H. Simpson's 1949 estimator of Simpson's Concentration is an unbiased estimator. It is not an unbiased Hill-Simpson estimator, but it is a pretty good one
}
