% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFuns.R
\name{subsam}
\alias{subsam}
\title{Take an abundance vector and subsample to size}
\usage{
subsam(ab_vec, size = sum(ab_vec))
}
\arguments{
\item{ab_vec}{A numeric vector of species abundances}

\item{size}{Number of individuals to sample, defaults to all of them, a scalar}
}
\value{
A numeric vector of species abundances, including 0's
}
\description{
Take a finite sample of individuals without replacement from a finite abundance vector.
}
\examples{
subsam(1:9, 15)
}
