% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot_gamma_and_lnorm.R
\name{try_SAD}
\alias{try_SAD}
\title{Relative abundances, given richness and distributional assumption}
\usage{
try_SAD(rich, x, distr)
}
\arguments{
\item{rich}{Integer, the total number of species in the species abundance
distribution.}

\item{x}{Non-negative scalar, the shape parameter for a gamma distribution.}

\item{distr}{Character string, one of \code{c("lnorm", "gamma")}}
}
\description{
Wrapper for \code{\link[stats]{qgamma}} and \code{\link[stats]{qlnorm}} that
takes the number of species and a shape parameter
and gives relative abundance estimates for each species.
}
\examples{
try_SAD(rich = 10, x = 0.2, distr = "lnorm")
try_SAD(rich = 10, x = 2, distr = "lnorm")

try_SAD(rich = 10, x = 0.2, distr = "gamma")
try_SAD(rich = 10, x = 2, distr = "gamma")

\dontrun{ try_SAD(rich = 10, x = -1, distr = "lnorm")
try_SAD(rich = 10, x = -1, distr = "gamma")
try_SAD(rich = 10, x = 0, distr = "gamma")
}

}
\seealso{
\code{\link[stats]{qgamma}, \link[stats]{qlnorm}, \link{fit_SAD}}
}
