% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{white_y}
\alias{white_y}
\title{Convenience function to blank y-axis elements for constructing multi-panel
plots.}
\usage{
white_y(p)
}
\arguments{
\item{p}{A ggplot.}
}
\value{
A ggplot.
}
\description{
This function whites out, but does not remove, y-axis elements for a
consistent plot size that can be included in a multi-panel plot
}
