% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beeAbunds}
\alias{beeAbunds}
\title{Bee specimens from 4 sites sampled with equal effort, summarized}
\format{
A tibble with 68 rows and 5 variables:
\itemize{
\item \emph{bee}: concatenated Latin binomial for bee species ID
\item \emph{\verb{Cold Soil_5}}: number of records of given species from site "Cold Soil"
during sampling round 5
\item \emph{\verb{Fox Hill_5}}: number of records of given species from site "Fox Hill"
during sampling round 5
\item \emph{\code{IAS_3}}: number of records of given species from site "IAS" during
sampling round 3
\item \emph{\verb{Lord Stirling_4}}: number of records of given species from site
"Lord Stirling" during sampling round 5
}
}
\usage{
beeAbunds
}
\description{
A dataset containing the number of specimens of each bee species
collected at each site during 15 30-minute transects.
}
\keyword{datasets}
