% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_funs.R
\name{Bootstrap.CI}
\alias{Bootstrap.CI}
\title{Bt_prob_abu(x) is a function of estimating the species probabilities in the bootstrap assemblage based on abundance data.}
\usage{
Bootstrap.CI(
  x,
  q,
  B = 1000,
  datatype = c("abundance", "incidence"),
  conf = 0.95
)
}
\arguments{
\item{x}{a vector of species sample frequencies (for abundance data) or incidence-based sample frequencies (1st entry must be the number of sampling unit).}

\item{q}{a numeric or a vector of diversity order.}

\item{B}{an integer to specify the number of replications in the bootstrap procedure, B = 1000 is suggested for constructing confidence intervals;
To save running time, use a smaller value (e.g. B = 200)..}

\item{datatype}{a character of data type,"abundance" or "incidence".}

\item{conf}{a confidence coefficient between 0 and 1.}
}
\value{
a numeric vector.
Bootstrap.CI(x,q,B,datatype,conf) is a function of calculating the bootstrapping standard error based on abundance data or incidence data.

a list, consisting of 3 matrices including respectively the difference between the average and lower confidence bound of the B bootstrap estimates,
the difference between the upper confidence bound and the average of the B bootstrap estimates, and the bootstrap standard error of the diversity estimate.
In each matrix, the first row gives the results for the empirical diversity, and the second row gives the results for the proposed diversity estimates.
Columns give the results for different orders of q.

@noRd
}
\description{
Bt_prob_abu(x) is a function of estimating the species probabilities in the bootstrap assemblage based on abundance data.
}
