% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSpadeR.R
\name{Bootstrap.CI_df}
\alias{Bootstrap.CI_df}
\title{Approximate CI for observed and asymptotic Hill diversity.}
\usage{
Bootstrap.CI_df(x, l, B = 1000, conf = 0.95)
}
\arguments{
\item{x}{a vector of species sample frequencies (for abundance data) or
incidence-based sample frequencies (1st entry must be the number of
sampling unit).}

\item{l}{Scaling exponent for the mean, can be any real number.}

\item{B}{an integer to specify the number of replications in the bootstrap
procedure, B = 1000 is suggested for constructing confidence intervals;
To save running time, use a smaller value (e.g. B = 200)..}

\item{conf}{a confidence coefficient between 0 and 1.}
}
\value{
a list, consisting of 3 matrices including respectively the
difference between the average and lower confidence bound of the B
bootstrap estimates, the difference between the upper confidence bound and
the average of the B bootstrap estimates, and the bootstrap standard error
of the diversity estimate. In each matrix, the first row gives the results
for the empirical diversity, and the second row gives the results for the
proposed diversity estimates. Columns give the results for different orders
of q.

@noRd
}
\description{
Functionally, a wrapper for \code{SpadeR:::Bootstrap_CI}, slimmed down to deal with
only abundance data and set to return a data.frame. This is the approximate CI suggested by Chao and Jost 2015 MEE. Source code copied and pasted from
\code{SpadeR:::Bootstrap_CI} from the R package SpadeR 0.1.1 by Anne Chao, K. H.
Ma, T. C. Hsieh and Chun-Huo Chiu
}
\references{
\insertRef{Chao2015}{MeanRarity}
}
