% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSpadeR.R
\name{Bt_prob_abu}
\alias{Bt_prob_abu}
\title{Generate "augmented" sample for bootstrapping}
\usage{
Bt_prob_abu(ab)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}
}
\value{
numeric vector of Estimated species frequencies
}
\description{
Estimates true species frequencies (including unobserved species). The
estimated species frequencies are bootstrapped to generate CI for diversity
estimates following the method of Chao and Jost 2015 MEE. We modified the
source code from \code{SpadeR} so that parameter names matched analogous use
elsewhere in the \code{MeanRarity} pacakge. This function is copied directly
from\code{SpadeR:::Chat.Ind()} from the R package SpadeR 0.1.1 by Anne Chao, K. H.
Ma, T. C. Hsieh and Chun-Huo Chiu.
}
\examples{

#a sample
ab <- c(1, 1, 1, 2, 2, 2, 3, 4, 5, 7, 9, 12, 13, 21, 25, 50)

#sample frequencies
freq <- (ab/sum(ab))
freq[order(freq, decreasing = TRUE)]

#estimated true frequencies
etf <- Bt_prob_abu(ab)
etf[order(etf, decreasing =TRUE)] # adjustment evident from 4th species on... 2 new rare species added
sum(etf) #rel. abundances still sum to 1

}
\references{
\insertRef{Chao2015}{MeanRarity}
}
