% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSpadeR.R
\name{Chao_Hill_abu}
\alias{Chao_Hill_abu}
\title{Estimate asymptotic Hill diversity}
\usage{
Chao_Hill_abu(ab, l, q = NULL)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{l}{Scaling exponent for the mean, can be any real number.}

\item{q}{Scalar, traditional Hill number scaling exponent, q = 1-l where l is the
scaling parameter for the generalized mean. Can be any real number.}
}
\value{
Scalar estimate of Hill diversity given \eqn{\ell}
}
\description{
Computes an estimate of the true community diversity based on a finite sample
assuming that individuals are randomly and independently sampled, after Chao
and Jost 2015. \code{SpadeR} source code modified to match our use of \eqn{\ell}
instead of \code{q} to parameterize Hill diversity. Code taken from
\code{SpadeR:::Chao_Hill_abu()}From the R package SpadeR 0.1.1 by Anne Chao, K. H.
Ma, T. C. Hsieh and Chun-Huo Chiu
}
\examples{

# generate sample
abs <- sample_infinite(fit_SAD(rich = 50, simpson = 20)[[3]], 150)

# estimate true Hill diversity
Chao_Hill_abu(abs, l = 1) # Chao1 estimate of richness lower bound
Chao_Hill_abu(abs, l = 0) # Asymptotic estimate of Hill-Shannon diversity
Chao_Hill_abu(abs, l = -1) # Asymptotic estimate of Hill-Simpson diversity

}
\references{
\insertRef{Chao2015}{MeanRarity}
}
