% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFuns.R
\name{dfun}
\alias{dfun}
\title{estimate Hill diversity}
\usage{
dfun(ab, l)
}
\arguments{
\item{ab}{a numeric vector of species abundances or relative abundances}

\item{l}{scaling exponent for the mean, can be any real number}
}
\value{

}
\description{
We parameterize Hill diversity \eqn{D} as a the frequency-weighted mean species rarity, with scaling exponent l
\deqn{D=\sum{p_i*r_i^{l}}^{-l}}
where rarity of species i \eqn{r_1=1/p_i}
}
\details{
This is equivalent to the "q" notation of Jost 2006
\deqn{D=\sum{p_i{q}}^{1-q}}
where \eqn{q=1-l}
}
