% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot_gamma_and_lnorm.R
\name{divers_gamma}
\alias{divers_gamma}
\title{Relative abundances given gamma disribution}
\usage{
divers_gamma(rich, x)
}
\arguments{
\item{rich}{Total number of species in the SAD, an integer}

\item{x}{Shape paramter for a gamma distribution, a scalar}
}
\description{
Wrapper for \code{[stats] qgamma} that
takes the number of species and a shape parameter
and gives relative abundance estimates for each species.
}
\examples{
divers_gamma(rich = 10, x = 0.2)

divers_gamma(rich = 10, x = 2)
}
\seealso{
qgamma
}
