% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot_gamma_and_lnorm.R
\name{fit_SAD}
\alias{fit_SAD}
\title{Fit a SAD given diversity and a distributional assumption}
\usage{
fit_SAD(
  rich = 50,
  simpson = 40,
  distr = "lnorm",
  int_lwr = 1e-04,
  int_uppr = 100,
  totAb = 1e+07,
  ...
)
}
\arguments{
\item{rich}{Total number of species in the SAD, an integer.}

\item{simpson}{Hill-Simpson diversity of the SAD, a real number in [1,rich].}

\item{distr}{Name of the distribution (\code{"lnorm"} or \code{"gamma"}).}

\item{int_lwr}{Lower bound of search space for \code{uniroot()}; default is a small
number close to 0 to deal with potential boundary issues for
\code{stats::uniroot} (a scalar).}

\item{int_uppr}{Upper bound of search space for \code{uniroot} (scalar).}

\item{totAb}{Not implemented; would be a necessary constraint for fitting
finite communities.}

\item{...}{pass arguments to other}
}
\value{
A list with three elements.

\code{distribution_info} Contains the name of the distribution and the fitted
shape parameter value.

\code{community_info} gives richness, Hill-Shannon, and Hill=Simpson
diversity of the SAD.

\code{rel_abundances} is a vector of relative abundances for each species
in SAD.
}
\description{
Takes a true richness, an inverse-Simpson diversity, and a distributional
assumption. Fits an optimal SAD given these constraints using
\code{stats::uniroot}.
}
\details{
The way the SAD is fit, we assume infinite abundance, but finite and
fixed diversity. In particular, richness is fixed. The parametric
fits use continuous distributions. Species abundances are given at
evenly spaced intervals along those continuous distributions. The distributions
are described by a shape parameter (the scale parameter of the lognormal
is fixed), and the value of that parameter is chosen such that the relative
abundances assigned to the species give the target Hill-Simpson diversity.
}
\examples{
#works
fit_SAD(distr = "gamma") #works
fit_SAD(distr = "lnorm", rich = 50, simpson = 2) #works
fit_SAD(distr = "gamma", rich = 50, simpson = 2) #works
fit_SAD(distr = "gamma", rich = 10, simpson = 6)

\dontrun{
fit_SAD(distr = "nonsense")  #returns error
fit_SAD(distr = "lnorm", rich = 50, simpson = 90) #returns error
}
}
\seealso{
\code{\link[stats]{uniroot}}, \code{\link{MeanRarity}}
}
