% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanRarity-package.R
\docType{package}
\name{MeanRarity-package}
\alias{MeanRarity}
\alias{MeanRarity-package}
\title{MeanRarity: Estimates Diversity Using The Mean Rarity Framework}
\description{
This package contains functions to estimate and visualize [Hill] diversity as a function of a scaling parameter \eqn{\ell}{"ell"}, as well as various convenience functions for sampling communities, estimating diversity from subsamples, and computing uncertainty estimates.
}
\author{
\strong{Maintainer}: Michael Roswell \email{mike.roswell@gmail.com} (\href{https://orcid.org/0000-0002-8479-9184}{ORCID})

Authors:
\itemize{
  \item Jonathan Dushoff \email{jdushoff@gmail.com} (\href{https://orcid.org/0000-0003-0506-4794}{ORCID})
}

}
\keyword{internal}
