% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{omit_y}
\alias{omit_y}
\title{Convenience function to omit y-axis elements for constructing multi-panel
plots.}
\usage{
omit_y(p)
}
\arguments{
\item{p}{A ggplot.}
}
\value{
A ggplot.
}
\description{
This function removes y-axis elements for plots intended to be
included in a multi-panel plot. WHen the y-axis is ommitted (rather than
whited-out as in \code{white_y}), the plotted area typically expands to fill
unused space.
}
\seealso{
\code{\link{white_y}}
}
