% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{power_trans}
\alias{power_trans}
\title{Power scale transformation}
\usage{
power_trans(pow, nb)
}
\arguments{
\item{pow}{Exponent of power transformation, scalar.}

\item{nb}{Number of desired breaks (approximate), scalar.}
}
\value{
A scale transformation object for plotting in ggplot.
}
\description{
This is a function for nice scale transformations for \code{ggplot2},
including when \code{pow < 0}.
}
\details{
This function is for power transformations (i.e. raising to a power and
inverse) This is useful for visualizing generalized means.
}
\seealso{
\code{\link[scales]{trans_new}}, \code{\link[scales]{trans_breaks}},
\code{\link{pfun}}, \code{\link{ipfun}}
}
