% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{prettify}
\alias{prettify}
\title{round numbers, more aggressively the larger they are}
\usage{
prettify(breaks)
}
\arguments{
\item{breaks}{is a numeric vector 
intended for the breaks returned by \code{scales::transbreaks}}
}
\value{
numeric vector of same length as input, rounded nicely
}
\description{
round numbers, more aggressively the larger they are
}
