% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFuns.R
\name{sample_finite}
\alias{sample_finite}
\title{Take an abundance vector and subsample to size}
\usage{
sample_finite(ab, size = sum(ab))
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{size}{Number of individuals to sample, defaults to all of them, a
scalar}
}
\value{
A numeric vector of species abundances, including 0's
}
\description{
Take a finite sample of individuals without replacement from an integer
abundance vector.
}
\examples{
sample_finite(1:9, 15)
}
\seealso{
\code{\link{sample_infinite}}
}
\concept{Sampling}
