% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFuns.R
\name{sample_infinite}
\alias{sample_infinite}
\title{Sample a [relative] abundance vector with replacement}
\usage{
sample_infinite(commSAD, size)
}
\arguments{
\item{commSAD}{Numeric vector of [relative] abundances.}

\item{size}{Scalar, number of individuals in finite samples.}
}
\value{
A vector of integer species abundances
}
\description{
Subsample the whole community with number of individuals = \code{size}
}
\seealso{
\code{\link{sample_finite}}
}
