% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFuns.R
\name{subcom}
\alias{subcom}
\title{Subsample of several community vectors
A wrapper of \code{\link{sample_finite}} to take a subset of a bunch of
communities, each subset of equal abundance.}
\usage{
subcom(ab, size)
}
\arguments{
\item{ab}{A numeric vector of species abundances or relative abundances.}

\item{size}{Number of individuals to sample, a scalar.}
}
\description{
Subsample of several community vectors
A wrapper of \code{\link{sample_finite}} to take a subset of a bunch of
communities, each subset of equal abundance.
}
\concept{Sampling}
