% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot_gamma_and_lnorm.R
\name{ur_distr}
\alias{ur_distr}
\title{Difference between target and realized diversity of simulated SAD (gamma
distribution).}
\usage{
ur_distr(x, rich = rich, simpson = simpson, distr = distr, totAb = totAb, ...)
}
\arguments{
\item{x}{Shape parameter for a either lognormal or gamma distribution, a scalar.}

\item{rich}{Total number of species in the SAD, an integer.}

\item{simpson}{Target value for inverse Simpson's concentration of the simulated
SAD, a scalar.}

\item{distr}{Distribution type (currently "\code{lnorm"} or \code{"gamma"}) to call for
"divers_" function, a character string.}

\item{...}{Additional arguments passed to other functions.}

\item{totAB}{Not implemented, could have a finite-size version with a fixed #
of individuals in pool.}
}
\value{
a scalar, the difference between empirical and target Hill-Simpson
diversity
}
\description{
Subtracts realized inverse simpson diversity of the simulated species
abundance distribution from the target value. When this difference = 0 the
shape parameter of the gamma distribution is considered optimal.
}
\examples{

ur_distr(x = 2, rich = 10, simpson = 6, distr = "gamma")
ur_distr(x = 0.2, rich = 10, simpson = 6, distr = "gamma")
ur_distr(x = 1.235383382 , rich = 10
     , simpson = 6, distr = "gamma") #very close, depends on sensitivity.
}
\seealso{
\code{\link{dfun}}
}
